+++ application/views/administrator_view.php
@@ -17,7 +17,8 @@
         </div>
         <div id="collapseOne" class="panel-collapse collapse in" role="tabpanel" aria-labelledby="headingOne">
             <div class="panel-body">
-
+                <a class="btn btn-success" role="button" href="/Administrator/newSupplies">Add New Supply</a>
+                <br><br>
                 <table class = "table table-bordered">
                     <tr>
                         <th>Item Id</th>
@@ -25,6 +26,7 @@
                         <th>On-Hand Quantity (g)</th>
                         <th>Receiving Quantity (g)</th>
                         <th>Price per Unit (cent)</th>
+                        <th>Action</th>
                     </tr>
                     {supplyItems}
                     <tr> <td><a class="btn btn-default" role="button" href="/Administrator/editSupplies/{id}">{id}</a></td>
@@ -32,6 +34,7 @@
                         <td>{qty_onhand}</td>
                         <td>{qty_inventory}</td>
                         <td>{price}</td>
+                        <td><a class="btn btn-primary" role="button" href="/Administrator/editSupplies/{id}">Edit</a> <a class="btn btn-danger" role="button" href="/Administrator/deleteSupply/{id}">Delete</a></td>
                     </tr>
                     {/supplyItems}
                 </table>
@@ -64,6 +67,7 @@
                         <th>Price</th>
                         <th>Units Available (Pieces)</th>
                         <th>Picture</th>
+                        <th>Action</th>
                     </tr>
                     {recipeItems}
                     <tr> <td><a class="btn btn-default" role="button" href="/Production/edit/{id}">{id}</a></td>
@@ -72,6 +76,7 @@
                         <td>{price}</td>
                         <td>{unit}</td>
                         <td><img class="scale-detail" src="/pix/{picture}"></td>
+                        <td><a class="btn btn-primary" role="button" href="/Production/edit/{id}">Edit</a> <a class="btn btn-danger" role="button" href="/Production/delete/{id}">Delete</a></td>
                     </tr>
                     {/recipeItems}
                 </table>